requirements.txt
requests
pydantic
websocket-client